physics.cpp cosf->std::cos for #52

+++ src/engine/world/physics.cpp
@@ -163,9 +163,9 @@ bool ellipsecollide(physent *d, const vec &dir, const vec &o, const vec &center,
           dangle = angle-d->yaw*RAD,
           eangle = angle-yaw*RAD,
           dx = d->xradius*cosf(dangle),
-          dy = d->yradius*sinf(dangle),
+          dy = d->yradius*std::sin(dangle),
           ex = xr*cosf(eangle),
-          ey = yr*sinf(eangle),
+          ey = yr*std::sin(eangle),
           dist = sqrtf(x*x + y*y) - sqrtf(dx*dx + dy*dy) - sqrtf(ex*ex + ey*ey);
     if(dist < 0)
     {
@@ -1305,7 +1305,7 @@ void vecfromyawpitch(float yaw, float pitch, int move, int strafe, vec &m)
 {
     if(move)
     {
-        m.x = move*-sinf(RAD*yaw);
+        m.x = move*-std::sin(RAD*yaw);
         m.y = move*cosf(RAD*yaw);
     }
     else
@@ -1317,7 +1317,7 @@ void vecfromyawpitch(float yaw, float pitch, int move, int strafe, vec &m)
     {
         m.x *= cosf(RAD*pitch);
         m.y *= cosf(RAD*pitch);
-        m.z = move*sinf(RAD*pitch);
+        m.z = move*std::sin(RAD*pitch);
     }
     else
     {
@@ -1327,7 +1327,7 @@ void vecfromyawpitch(float yaw, float pitch, int move, int strafe, vec &m)
     if(strafe)
     {
         m.x += strafe*cosf(RAD*yaw);
-        m.y += strafe*sinf(RAD*yaw);
+        m.y += strafe*std::sin(RAD*yaw);
     }
 }
